% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{doypa}
\alias{doypa}
\title{Get a DOYPAColors color palette}
\usage{
doypa(
  palette = NULL,
  n = NULL,
  reverse = FALSE,
  gradient = FALSE,
  type = "all",
  colorblind = FALSE
)
}
\arguments{
\item{palette}{Character string specifying the desired palette name.}

\item{n}{Number of colors needed. Must be a positive numeric value.}

\item{reverse}{Logical. If 'TRUE', reverses the order of colors in the palette.}

\item{gradient}{Logical. If 'TRUE', interpolates a color gradient of n colors between the first and last colors of the palette. If 'FALSE', returns the first n colors of the palette.}

\item{type}{Character specifying the type of palettes to consider when choosing a random one.
Options: "all" (default), "seq" (sequential), "div" (diverging), "qual" (qualitative).}

\item{colorblind}{Logical. Set to 'TRUE' for palettes designed with consideration for colorblind individuals.}
}
\value{
A vector of 'n' colors.
}
\description{
This function generates a color vector of n colors for a specified palette.
}
\examples{
# Generate a random color palette
color_palette <- doypa()
preview_pal(colors = color_palette)

# Generate a color palette of 5 colors using the "buzz" palette
color_palette <- doypa(palette = "buzz", n = 5)
preview_pal(colors = color_palette)

}
\seealso{
'list_doypa_pals' to list available palettes.
'preview_doypa_pals' to preview all color palettes.
}
