% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doypa.R
\docType{data}
\name{doypa_palettes}
\alias{doypa_palettes}
\title{doypa color palettes}
\format{
An object of class \code{list} of length 11.
}
\usage{
doypa_palettes
}
\description{
doypa color palettes
}
\keyword{datasets}
