% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOYPAColors_functions.R
\name{list_doypa_pals}
\alias{list_doypa_pals}
\title{List available DOYPAColors color palettes}
\usage{
list_doypa_pals()
}
\value{
A character vector containing the names of available color palettes.
}
\description{
This function provides a list of all available DOYPAColors color palettes.
}
\examples{
# List available DOYPAColors color palettes
palette_names <- list_doypa_pals()
print(palette_names)
}
