% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_doypa}
\alias{scale_colour_doypa}
\title{Color scale for ggplot2 with DOYPAColors color palettes}
\usage{
scale_colour_doypa(
  palette = NULL,
  n = NULL,
  reverse = FALSE,
  gradient = FALSE,
  discrete = FALSE,
  type = "all",
  colorblind = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character string specifying the desired palette name.}

\item{n}{Number of colors needed.}

\item{reverse}{Boolean indicating whether to reverse the order of colors in the palette (default: FALSE).}

\item{gradient}{Boolean indicating whether to use a color gradient (interpolation) or a vector of colors (default: FALSE).}

\item{discrete}{Boolean indicating whether to generate a discrete or continuous palette (default: continuous).}

\item{type}{Character specifying the type of palettes to consider when choosing a random one.
Options: "all" (default), "seq" (sequential), "div" (diverging), "qual" (qualitative).}

\item{colorblind}{Logical. Set to 'TRUE' for palettes designed with consideration for colorblind individuals.}

\item{...}{Additional parameters to pass to ggplot2's scale_colour functions.}
}
\value{
A `ggplot2` color scale suitable for adding to a `ggplot2` object to control color aesthetics.
}
\description{
This function sets the color scale for ggplot2 using DOYPAColors color palettes.
}
