% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doypa.R
\name{doypa}
\alias{doypa}
\title{doypa color palette generator}
\usage{
doypa(palette = "main", ...)
}
\arguments{
\item{palette}{Character string of the desired palette:
\itemize{
 \item "main"
 \item "tidy"
 \item "autumn"
 \item "forest"
 \item "ocean"
 \item "magma"
 \item "neon"
 \item "luxury"
 \item "santapola"
 \item "chinatown"
 \item "bahamas"
}}

\item{...}{Additional parameters of grDevices::colorRampPalette}
}
\description{
This is a function to turn the default color schemes
into a compatible color palette with ggplot.
}
\details{
See \code{\link[doypa]{get_doypa_palettes}} for detailed
information of the color palettes.`
}
