% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{doypa_palette}
\alias{doypa_palette}
\title{Create a DOYPAColors color palette function}
\usage{
doypa_palette(
  palette = NULL,
  reverse = FALSE,
  gradient = FALSE,
  type = "all",
  colorblind = FALSE
)
}
\arguments{
\item{palette}{Character string specifying the desired palette name.}

\item{reverse}{Logical. If 'TRUE', reverses the order of colors in the palette.}

\item{gradient}{Logical. If 'TRUE', the returned function generates a color gradient (interpolation) instead of a vector of colors.}

\item{type}{Character specifying the type of palettes to consider when choosing a random one.
Options: "all" (default), "seq" (sequential), "div" (diverging), "qual" (qualitative).}

\item{colorblind}{Logical. Set to 'TRUE' for palettes designed with consideration for colorblind individuals.}
}
\value{
A function that generates a vector of 'n' colors or a color gradient.
}
\description{
This function creates a function that generates a color vector of n colors 
for a specified DOYPAColors palette.
}
\examples{
\dontrun{
# Create a function for generating colors from the "buzz" palette
get_buzz_palette <- doypa_palette(palette = "buzz")
color_palette <- get_buzz_palette(5)
preview_pal(colors = color_palette)
}
}
\keyword{internal}
