% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preview_pal}
\alias{preview_pal}
\title{Preview a Single Color Palette}
\usage{
preview_pal(colors = NULL)
}
\arguments{
\item{colors}{A vector of colors to preview.}
}
\value{
A 'ggplot' visualization of a color palette for exploring and selecting color schemes in your data visualizations.
}
\description{
This function generates a visual preview of a single color palette, allowing you to explore and select color schemes for your data visualizations.
}
\examples{
# Preview a custom vector of colors
custom_colors <- c("#FF5733", "#33FF57", "#5733FF")
preview_pal(colors = custom_colors)

# Preview default DOYPAColors palettes
preview_pal(doypa("buzz"))
}
