% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_doypa}
\alias{scale_color_doypa}
\title{Color scale for ggplot2 with DOYPAColors color palettes}
\usage{
scale_color_doypa(
  palette = NULL,
  n = NULL,
  reverse = FALSE,
  gradient = FALSE,
  discrete = FALSE,
  type = "all",
  colorblind = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character string specifying the desired palette name.}

\item{n}{Number of colors needed.}

\item{reverse}{Boolean indicating whether to reverse the order of colors in the palette (default: FALSE).}

\item{gradient}{Boolean indicating whether to use a color gradient (interpolation) or a vector of colors (default: FALSE).}

\item{discrete}{Boolean indicating whether to generate a discrete or continuous palette (default: continuous).}

\item{type}{Character specifying the type of palettes to consider when choosing a random one.
Options: "all" (default), "seq" (sequential), "div" (diverging), "qual" (qualitative).}

\item{colorblind}{Logical. Set to 'TRUE' for palettes designed with consideration for colorblind individuals.}

\item{...}{Additional parameters to pass to ggplot2's scale_color functions.}
}
\value{
A `ggplot2` color scale suitable for adding to a `ggplot2` object to control color aesthetics.
}
\description{
This function sets the color scale for ggplot2 using DOYPAColors color palettes.
}
\examples{
library(ggplot2)

# Discrete data
data(iris)
disc <- ggplot(iris, aes(x = Petal.Width, y = Petal.Length, color = Species)) +
  geom_point() + theme_classic()
disc <- disc + scale_color_doypa(palette = "buzz", discrete = TRUE)
print(disc)

# Continuous data
cont <- ggplot(iris, aes(x = Petal.Width, y = Petal.Length, color = Sepal.Length)) +
  geom_point() + theme_classic()
cont <- cont + scale_color_doypa(palette = "buzz")
print(cont)
}
