% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_doypa}
\alias{scale_colour_doypa}
\title{Color scale for ggplot2 with DOYPAColors color palettes}
\usage{
scale_colour_doypa(
  palette = NULL,
  n = NULL,
  reverse = FALSE,
  gradient = FALSE,
  discrete = FALSE,
  type = "all",
  colorblind = FALSE,
  ...
)
}
\arguments{
\item{palette}{A character string specifying the name of the palette to use. If `NULL`, a random palette will be selected.}

\item{n}{Number of colors needed. If `NULL`, it will default to the number of levels for discrete scales or to a continuous gradient for continuous scales.}

\item{reverse}{A logical value indicating whether to reverse the order of colors in the palette. Default is `FALSE`.}

\item{gradient}{A logical value indicating whether to interpolate colors as a  gradient of `n` colors between the first and last colors of the palette. If `FALSE`, returns the first `n` colors of the palette.}

\item{discrete}{Boolean indicating whether to generate a discrete or continuous palette (default: continuous).}

\item{type}{A character string specifying the type of palettes to select from: "all", "seq" (sequential), "div" (diverging), or "qual" (qualitative). Default is "all".}

\item{colorblind}{A logical value indicating whether to restrict the palette to colorblind-friendly options. Default is `FALSE`.}

\item{...}{Additional parameters passed to ggplot2's `scale_colour_*` functions.}
}
\value{
A `ggplot2` color scale suitable for adding to a `ggplot2` object to control color aesthetics.
}
\description{
This function sets the color scale for ggplot2 using DOYPAColors color palettes, allowing you to 
apply these palettes to both discrete and continuous data.
}
