% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOYPAColors_functions.R
\name{scale_colour_doypa}
\alias{scale_colour_doypa}
\title{Color scale for ggplot2 with DOYPAColors color palettes}
\usage{
scale_colour_doypa(
  palette = NULL,
  n = NULL,
  reverse = FALSE,
  ramp = FALSE,
  discrete = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character string specifying the desired palette name.
Available palettes: list_doypa_pals().}

\item{n}{Number of colors needed.}

\item{reverse}{Boolean indicating whether to reverse the order of colors in the palette (default: FALSE).}

\item{ramp}{Boolean indicating whether to use a color ramp (gradient) or a vector of colors (default: FALSE).}

\item{discrete}{Boolean indicating whether to generate a discrete or continuous palette (default: continuous).}

\item{...}{Additional parameters to pass to ggplot2's scale_color functions.}
}
\description{
This function sets the color scale for ggplot2 using DOYPAColors color palettes.
}
