% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doypa.R
\name{scale_fill_doypa}
\alias{scale_fill_doypa}
\title{Fill scale for ggplot2}
\usage{
scale_fill_doypa(palette = "main", discrete = FALSE, ...)
}
\arguments{
\item{palette}{Character string of the desired palette:
\itemize{
 \item "main"
 \item "tidy"
 \item "autumn"
 \item "forest"
 \item "ocean"
 \item "magma"
 \item "neon"
 \item "luxury"
 \item "santapola"
 \item "chinatown"
 \item "bahamas"
}}

\item{discrete}{Boolean indicating whether generating a
discrete or continuous palette (default: generate continuous \code{FALSE}).}

\item{...}{Parameters of \code{\link[ggplot2]{discrete_scale}} or
\code{\link[ggplot2]{scale_fill_gradientn}}, deppending if discrete
is TRUE or FALSE.}
}
\description{
Scale fill function for \code{\link[ggplot2]{ggplot2}}.
}
\examples{
library(doypa)
library(ggplot2)
gg <- ggplot(PlantGrowth, aes(x=weight)) + geom_histogram()
gg <- gg + scale_fill_doypa(palette = "main")

}
